Move CLI related methods to separate file
